feat: Roadmap

+++ apps/web/pages/api/pages/settings/remove-domain.ts
@@ -1,28 +1,47 @@
-import { NextApiRequest, NextApiResponse } from "next";
-import { getSupabaseServerClient } from "../../../../utils/supabase/supabase-admin";
+import { supabaseAdmin } from "@changes-page/supabase/admin";
+import isFQDN from "validator/lib/isFQDN";
+import { withAuth } from "../../../../utils/withAuth";
 
-async function removeDomain(req: NextApiRequest, res: NextApiResponse) {
-  const { user } = await getSupabaseServerClient({ req, res });
+const removeDomain = withAuth<{ success: boolean }>(
+  async (req, res, { user }) => {
+    const { domain } = req.body;
 
-  const { domain } = req.body;
+    console.log("removeDomain", user?.id, `domain: ${domain}`);
 
-  console.log("removeDomain", user?.id, `domain: ${domain}`);
-
-  const response = await fetch(
-    `https://api.vercel.com/v8/projects/${process.env.VERCEL_PAGES_PROJECT_ID}/domains/${domain}?teamId=${process.env.VERCEL_TEAM_ID}`,
-    {
-      headers: {
-        Authorization: `Bearer ${process.env.VERCEL_AUTH_TOKEN}`,
-      },
-      method: "DELETE",
+    if (
+      typeof domain !== "string" ||
+      !isFQDN(domain, { require_tld: true, allow_underscores: false })
+    ) {
+      return res.status(400).json({ success: false });
     }
-  );
 
-  await response.json();
+    // check custom domain ownership
+    const { data, error } = await supabaseAdmin
+      .from("page_settings")
+      .select("page_id,custom_domain, pages(user_id)")
+      .eq("custom_domain", domain)
+      .single();
+    if (!data || error || data?.pages?.user_id !== user.id) {
+      return res.status(400).json({ success: false });
+    }
 
-  res.status(200).json({
-    success: true,
-  });
-}
+    const safeDomain = data.custom_domain;
+    const response = await fetch(
+      `https://api.vercel.com/v8/projects/${process.env.VERCEL_PAGES_PROJECT_ID}/domains/${safeDomain}?teamId=${process.env.VERCEL_TEAM_ID}`,
+      {
+        headers: {
+          Authorization: `Bearer ${process.env.VERCEL_AUTH_TOKEN}`,
+        },
+        method: "DELETE",
+      }
+    );
+
+    await response.json();
+
+    res.status(200).json({
+      success: true,
+    });
+  }
+);
 
 export default removeDomain;
